+++ lib/philiprehberger/graph.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'json'
 require 'set'
 
 require_relative 'graph/version'